SmokeTest in FileCheck umbenannt und eigene Stage auf Travis CI konfiguriert, die ohne WordPress auskommt
