+++ .travis.yml
@@ -18,12 +18,24 @@ env:
   - WP_VERSION=4.9.8 WP_MULTISITE=0
   - WP_VERSION=nightly WP_MULTISITE=0
 
+stages:
+  - validate
+  - test
+
+jobs:
+  include:
+  - stage: validate
+    name: "Basic Checks"
+    php: 7.2
+    install: composer require bartlett/php-compatinfo
+    script: phpunit -c phpunit-checks.xml --no-coverage
+
 matrix:
   allow_failures:
     - env: WP_VERSION=nightly WP_MULTISITE=0
   fast_finish: true
 
-before_script:
+install:
   - bash bin/install-wp-tests.sh wordpress_test root '' localhost $WP_VERSION
   - composer install --prefer-source
